+++ packages/cli/src/commands/generate.ts
@@ -174,6 +174,7 @@ async function outputCodegen(
   whitelist: string[] | null,
 ) {
   const {
+    commonFileContent,
     descriptorsFileContent,
     descriptorTypesFiles,
     metadataTypes,
@@ -186,6 +187,7 @@ async function outputCodegen(
       metadataTypes: "./metadataTypes",
       types: "./common-types",
       descriptorValues: "./descriptors",
+      common: "./common",
     },
     {
       whitelist: whitelist ?? undefined,
@@ -204,6 +206,7 @@ async function outputCodegen(
 
   await fs.mkdir(outputFolder, { recursive: true })
 
+  await fs.writeFile(path.join(outputFolder, "common.ts"), commonFileContent)
   // Going through base64 conversion instead of using binary loader because of esbuild issue
   // https://github.com/evanw/esbuild/issues/3894
   const metadataTypesBase64 = Buffer.from(