+++ packages/codegen/src/generate-descriptors.ts
@@ -56,6 +56,7 @@ export const generateDescriptors = (
     metadataTypes: string
     types: string
     descriptorValues: string
+    common: string
   },
   genesis?: string,
 ) => {
@@ -358,23 +359,7 @@ export const generateDescriptors = (
   // https://github.com/evanw/esbuild/issues/3894
   const imports = `import {${clientImports.join(", ")}} from "${paths.client}";
   import {${commonTypeImports.join(", ")}} from "${paths.types}";
-
-  const toBinary = (() => {
-    const table = new Uint8Array(128);
-    for (let i = 0; i < 64; i++) table[i < 26 ? i + 65 : i < 52 ? i + 71 : i < 62 ? i - 4 : i * 4 - 205] = i;
-    return (base64: string) => {
-      const n = base64.length,
-        bytes = new Uint8Array((n - Number(base64[n - 1] === '=') - Number(base64[n - 2] === '=')) * 3 / 4 | 0);
-      for (let i2 = 0, j = 0; i2 < n;) {
-        const c0 = table[base64.charCodeAt(i2++)], c1 = table[base64.charCodeAt(i2++)];
-        const c2 = table[base64.charCodeAt(i2++)], c3 = table[base64.charCodeAt(i2++)];
-        bytes[j++] = c0 << 2 | c1 >> 4;
-        bytes[j++] = c1 << 4 | c2 >> 2;
-        bytes[j++] = c2 << 6 | c3;
-      }
-      return bytes;
-    };
-  })();
+  import { toBinary } from "${paths.common}"
 
   const descriptorValues = import("${paths.descriptorValues}").then(module => module["${prefix}"]);
   const metadataTypes = import("${paths.metadataTypes}").then(